final submission of maxpairwise product

+++ maxpairwiseproduct.cpp
@@ -2,18 +2,18 @@
 #include <vector>
 #include <algorithm>
 
-long long MaxPairwiseProduct(const std::vector<int> &numbers)
+long long MaxPairwiseProduct(const std::vector<long long> &numbers)
 {
     long long max_product = 0;
-    int n = numbers.size();
-    int max_ind1 = -1;
-    for (int i = 0; i < n; ++i)
+    long long n = numbers.size();
+    long long max_ind1 = -1;
+    for (long long i = 0; i < n; ++i)
     {
         if ((max_ind1 == -1) || (numbers[i] > numbers[max_ind1]))
             max_ind1 = i;
     }
-    int max_ind2 = -1;
-    for (int j = 0; j < n; ++j)
+    long long max_ind2 = -1;
+    for (long long j = 0; j < n; ++j)
     {
         if ((j != max_ind1) && ((max_ind2 == -1) || (numbers[j] > numbers[max_ind2])))
             max_ind2 = j;
@@ -28,7 +28,7 @@ int main()
 {
     int n;
     std::cin >> n;
-    std::vector<int> numbers(n);
+    std::vector<long long> numbers(n);
     for (int i = 0; i < n; ++i)
     {
         std::cin >> numbers[i];
